+++ ProjectGE/src/ProjectGE/Rendering/DirectX12/RenderComponents/DirectX12IndexBuffer.cpp
@@ -2,7 +2,7 @@
 #include "DirectX12IndexBuffer.h"
 
 namespace ProjectGE {
-	DirectX12IndexBuffer::DirectX12IndexBuffer(unsigned short* buffer, unsigned int count) : m_Buffer(Scope<DirectX12Buffer<unsigned short>>(new DirectX12Buffer<unsigned short>(buffer, count, "Index Buffer"))), m_Count(count) {}
+	DirectX12IndexBuffer::DirectX12IndexBuffer(unsigned int* buffer, unsigned int count) : m_Buffer(Scope<DirectX12Buffer<unsigned int>>(new DirectX12Buffer<unsigned int>(buffer, count, "Index Buffer"))), m_Count(count) {}
 
 	D3D12_INDEX_BUFFER_VIEW DirectX12IndexBuffer::GenerateView()
 	{
